requirements/requirements.txt
apache-airflow-providers-google==8.11.0
# google-api-python-client
# google-auth-httplib2
# google-auth-oauthlib
# google-cloud-language
gcsfs==2022.5.0
pandas==1.5.3
requests==2.27.1
python-dotenv==0.20.0
